+++ CHANGELOG.md
@@ -18,6 +18,7 @@ singularity on pypi), and the versions here will coincide with these releases.
 
 ## [master](https://github.com/singularityhub/singularity-cli/tree/master)
  - adding testing for 3.1.0 with Singularity Orbs (0.0.53)
+ - inspect returns parsed json on success, or full message / return code otherwise
  - instance stop all missing check for Singularity V3.+ (0.0.52)
  - fixing bug with instances list, name not taken into account (0.0.51)
  - additional of args to instance start commands (0.0.50)